+++ src/tests/Morphant.Generator.IntegrationTests.CSharp9/Scenarios/ByConvention_7ef0668e/Applies_selection_to_ByConvention_but_not_explicit_Construct.cs
@@ -34,6 +34,10 @@ namespace Morphant.Generator.IntegrationTests.CSharp9.Scenarios.ByConvention_7ef
         public ExplicitByConventionDestination()
         {
         }
+
+        public ExplicitByConventionDestination(int id)
+        {
+        }
     }
 
     public sealed class ExplicitConstructDestination